WIP: Untested but fully developed code; incomplete doc added
