+++ apps/docs/src/pages/en/communities/grant-access-to-additional-products.md
@@ -1,0 +1,21 @@
+---
+title: Grant Access to Additional Products for Members
+description: Guide to grant access to additional products for community members
+layout: ../../../layouts/MainLayout.astro
+---
+
+## Steps to Grant Access to Additional Products
+
+1. Navigate to the **Memberships** section in your community dashboard.
+2. Click on the member's name/email you want to manage.
+3. You will be taken to user's edit screen.
+
+## FAQ
+
+1.  When a members gets access to additional products, are these enrollments be tracked?
+    Yes, these enrollments are tracked and will show up in the product's enrollment metric. However, they will not count towards the product's sales metrics.
+
+2.  Will the included product's email sequences be triggered?
+    Yes, the included product's email sequences will be triggered.
+
+        ![User edit screen](/assets/communities/member-permissions.png)
