+++ apps/web/components/public/payments/checkout.tsx
@@ -1,11 +1,10 @@
 "use client";
 
-import Image from "next/image";
 import { useContext, useEffect, useState } from "react";
 import { useForm, FormProvider } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import * as z from "zod";
-import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group";
+import { RadioGroup } from "@/components/ui/radio-group";
 import {
     FormControl,
     FormField,
@@ -13,12 +12,7 @@ import {
     FormLabel,
     FormMessage,
 } from "@/components/ui/form";
-import {
-    Collapsible,
-    CollapsibleContent,
-    CollapsibleTrigger,
-} from "@/components/ui/collapsible";
-import { Check, ChevronUp, ShoppingCart, X, Star, Package } from "lucide-react";
+import { Check, X } from "lucide-react";
 import { LoginForm } from "./login-form";
 import {
     PaymentPlan,
@@ -38,359 +32,11 @@ import { FetchBuilder } from "@courselit/utils";
 import { useRouter } from "next/navigation";
 import { loadStripe } from "@stripe/stripe-js";
 import { getSymbolFromCurrency, useToast } from "@courselit/components-library";
-import { getPlanPrice } from "@ui-lib/utils";
 import Script from "next/script";
-import {
-    Button,
-    Header3,
-    Header4,
-    PageCard,
-    PageCardContent,
-    Text1,
-} from "@courselit/page-primitives";
-import { CHECKOUT_PAGE_ORDER_SUMMARY } from "@ui-config/strings";
+import { Button, Header3, Text1 } from "@courselit/page-primitives";
+import { PaymentPlanCard } from "./payment-plan-card";
+import { MobileOrderSummary, DesktopOrderSummary } from "./order-summary";
 const { PaymentPlanType: paymentPlanType } = Constants;
-
-interface PaymentPlanCardProps {
-    plan: PaymentPlan;
-    isSelected: boolean;
-    isRecommended: boolean;
-    isLoggedIn: boolean;
-    currencySymbol: string;
-    includedProducts: Course[];
-    theme: any;
-    onSelect: (planId: string) => void;
-}
-
-function PaymentPlanCard({
-    plan,
-    isSelected,
-    isRecommended,
-    isLoggedIn,
-    currencySymbol,
-    includedProducts,
-    theme,
-    onSelect,
-}: PaymentPlanCardProps) {
-    const planPrice = getPlanPrice(plan);
-    const planIncludedProducts = getIncludedProductsDescription(
-        plan,
-        includedProducts,
-    );
-
-    return (
-        <FormItem className="space-y-0">
-            <FormControl>
-                <PageCard
-                    theme={theme.theme}
-                    className={`relative transition-all ${
-                        isSelected
-                            ? "border-primary ring-2 ring-primary/20"
-                            : isLoggedIn
-                              ? "hover:border-muted-foreground/50"
-                              : ""
-                    } ${!isLoggedIn ? "opacity-50 cursor-not-allowed" : "cursor-pointer"}`}
-                    onClick={() => {
-                        if (isLoggedIn) {
-                            onSelect(plan.planId);
-                        }
-                    }}
-                >
-                    <PageCardContent theme={theme.theme}>
-                        {/* Recommended Badge */}
-                        {isRecommended && (
-                            <div className="absolute -top-3 left-6">
-                                <div className="bg-primary text-primary-foreground px-3 py-1 rounded-full text-xs font-medium flex items-center gap-1">
-                                    <Star className="w-3 h-3" />
-                                    Recommended
-                                </div>
-                            </div>
-                        )}
-
-                        {/* Radio Button and Plan Header */}
-                        <div className="flex items-start justify-between mb-4">
-                            <div className="flex items-center gap-3">
-                                <RadioGroupItem
-                                    value={plan.planId}
-                                    disabled={!isLoggedIn}
-                                    className="mt-1"
-                                />
-                                <div>
-                                    <FormLabel className="text-lg font-semibold cursor-pointer">
-                                        {plan.name}
-                                    </FormLabel>
-                                </div>
-                            </div>
-                        </div>
-
-                        {/* Pricing */}
-                        <div className="mb-4">
-                            <div className="flex items-baseline gap-1">
-                                <span className="text-2xl font-bold text-foreground">
-                                    {currencySymbol}
-                                    {planPrice.amount.toFixed(2)}
-                                </span>
-                                {planPrice.period && (
-                                    <span className="text-muted-foreground text-sm">
-                                        {planPrice.period}
-                                    </span>
-                                )}
-                            </div>
-                            {plan.type === paymentPlanType.ONE_TIME && (
-                                <span className="text-sm text-muted-foreground">
-                                    one-time
-                                </span>
-                            )}
-                        </div>
-
-                        {/* Features */}
-                        <div className="space-y-3">
-                            {/* Included Products Count */}
-                            {planIncludedProducts.length > 0 && (
-                                <div className="flex items-center gap-2 text-sm text-foreground">
-                                    <Package className="w-4 h-4 text-muted-foreground" />
-                                    <span>
-                                        {planIncludedProducts.length} products
-                                        included
-                                    </span>
-                                </div>
-                            )}
-
-                            {/* Real included product chips */}
-                            {planIncludedProducts.length > 0 && (
-                                <div className="flex flex-wrap gap-2">
-                                    {planIncludedProducts.map((product) => (
-                                        <span
-                                            key={product.courseId}
-                                            className="px-3 py-1 border border-border bg-background text-foreground rounded text-sm"
-                                        >
-                                            {product.title}
-                                        </span>
-                                    ))}
-                                </div>
-                            )}
-                        </div>
-
-                        {/* Description */}
-                        {plan.description && (
-                            <div className="mt-4">
-                                <Text1
-                                    theme={theme.theme}
-                                    className="text-sm text-muted-foreground whitespace-pre-wrap"
-                                >
-                                    {plan.description}
-                                </Text1>
-                            </div>
-                        )}
-                    </PageCardContent>
-                </PageCard>
-            </FormControl>
-        </FormItem>
-    );
-}
-
-interface OrderSummaryProps {
-    product: Product;
-    selectedPlan: PaymentPlan | null;
-    paymentPlans: PaymentPlan[];
-    currencySymbol: string;
-    theme: any;
-    isOrderSummaryOpen: boolean;
-    setIsOrderSummaryOpen: (open: boolean) => void;
-}
-
-function MobileOrderSummary({
-    product,
-    selectedPlan,
-    paymentPlans,
-    currencySymbol,
-    theme,
-    isOrderSummaryOpen,
-    setIsOrderSummaryOpen,
-}: OrderSummaryProps) {
-    return (
-        <div className="md:hidden w-full sticky top-20 z-11 mb-8">
-            <PageCard theme={theme.theme}>
-                <PageCardContent theme={theme.theme} className="p-0">
-                    <div className="w-full">
-                        <div
-                            className="flex items-center justify-between p-4 border-b w-full"
-                            style={{
-                                borderBottomColor: theme.theme.colors.border,
-                            }}
-                        >
-                            <div className="flex items-center gap-1">
-                                <ShoppingCart className="h-5 w-5" />
-                                <Text1
-                                    className="p-0 !m-0 h-auto font-normal hover:bg-transparent flex items-center"
-                                    onClick={() =>
-                                        setIsOrderSummaryOpen(
-                                            !isOrderSummaryOpen,
-                                        )
-                                    }
-                                    theme={theme.theme}
-                                >
-                                    {isOrderSummaryOpen ? "Hide" : "Show"} order
-                                    summary
-                                    <ChevronUp
-                                        className={`h-4 w-4 ml-1 transition-transform duration-200 ${isOrderSummaryOpen ? "" : "rotate-180"}`}
-                                    />
-                                </Text1>
-                            </div>
-                            <div className="font-medium flex items-center">
-                                {currencySymbol}
-                                {getPlanPrice(
-                                    selectedPlan || paymentPlans[0],
-                                ).amount.toFixed(2)}
-                                <span className="text-sm text-muted-foreground ml-1">
-                                    {
-                                        getPlanPrice(
-                                            selectedPlan || paymentPlans[0],
-                                        ).period
-                                    }
-                                </span>
-                            </div>
-                        </div>
-                    </div>
-
-                    <Collapsible
-                        open={isOrderSummaryOpen}
-                        onOpenChange={setIsOrderSummaryOpen}
-                    >
-                        <CollapsibleTrigger className="sr-only">
-                            Toggle order summary
-                        </CollapsibleTrigger>
-                        <CollapsibleContent
-                            style={{
-                                borderBottomColor: theme.theme.colors.border,
-                            }}
-                        >
-                            <div className="p-4 space-y-4">
-                                <div className="flex gap-4">
-                                    <div className="h-16 w-16 relative rounded-lg overflow-hidden">
-                                        <Image
-                                            src={
-                                                product.featuredImage ||
-                                                "/courselit_backdrop_square.webp"
-                                            }
-                                            alt={product.name}
-                                            fill
-                                            className="object-cover"
-                                        />
-                                    </div>
-                                    <div>
-                                        <Header3 theme={theme.theme}>
-                                            {product.name}
-                                        </Header3>
-                                        {product.description && (
-                                            <Text1 theme={theme.theme}>
-                                                {product.description}
-                                            </Text1>
-                                        )}
-                                    </div>
-                                </div>
-
-                                {selectedPlan && (
-                                    <div className="space-y-4">
-                                        <div
-                                            className="flex justify-between pt-4 border-t"
-                                            style={{
-                                                borderTopColor:
-                                                    theme.theme.colors.border,
-                                            }}
-                                        >
-                                            <Header3 theme={theme.theme}>
-                                                Total
-                                            </Header3>
-                                            <div className="text-right flex items-center">
-                                                <Header4 className="font-medium">
-                                                    {currencySymbol}
-                                                    {getPlanPrice(
-                                                        selectedPlan,
-                                                    ).amount.toFixed(2)}
-                                                </Header4>
-                                                <Text1 theme={theme.theme}>
-                                                    {
-                                                        getPlanPrice(
-                                                            selectedPlan,
-                                                        ).period
-                                                    }
-                                                </Text1>
-                                            </div>
-                                        </div>
-                                    </div>
-                                )}
-                            </div>
-                        </CollapsibleContent>
-                    </Collapsible>
-                </PageCardContent>
-            </PageCard>
-        </div>
-    );
-}
-
-function DesktopOrderSummary({
-    product,
-    selectedPlan,
-    currencySymbol,
-    theme,
-}: Omit<
-    OrderSummaryProps,
-    "paymentPlans" | "isOrderSummaryOpen" | "setIsOrderSummaryOpen"
->) {
-    return (
-        <PageCard theme={theme.theme}>
-            <PageCardContent theme={theme.theme} className="space-y-4">
-                <Header3 theme={theme.theme}>
-                    {CHECKOUT_PAGE_ORDER_SUMMARY}
-                </Header3>
-                <div className="flex items-start gap-4 pb-4">
-                    <div className="h-16 w-16 relative rounded-lg overflow-hidden bg-gray-100">
-                        <Image
-                            src={
-                                product.featuredImage ||
-                                "/courselit_backdrop_square.webp"
-                            }
-                            alt={product.name}
-                            fill
-                            className="object-cover"
-                        />
-                    </div>
-                    <div>
-                        <Header4 theme={theme.theme}>{product.name}</Header4>
-                        {product.description && (
-                            <Text1 theme={theme.theme}>
-                                {product.description}
-                            </Text1>
-                        )}
-                    </div>
-                </div>
-                {selectedPlan && (
-                    <div
-                        className="mt-4 pt-4 border-t"
-                        style={{
-                            borderTopColor: theme.theme.colors.border,
-                        }}
-                    >
-                        <div className="flex justify-between items-center">
-                            <Header4 theme={theme.theme}>Total</Header4>
-                            <Header4 theme={theme.theme}>
-                                {currencySymbol}
-                                {getPlanPrice(selectedPlan).amount.toFixed(2)}
-                                <span className="text-sm text-muted-foreground ml-1">
-                                    {getPlanPrice(selectedPlan).period}
-                                </span>
-                            </Header4>
-                        </div>
-                        <Text1 theme={theme.theme}>
-                            {getPlanDescription(selectedPlan, currencySymbol)}
-                        </Text1>
-                    </div>
-                )}
-            </PageCardContent>
-        </PageCard>
-    );
-}
 
 export interface Product {
     id: string;
@@ -414,46 +60,6 @@ const formSchema = z.object({
     selectedPlan: z.string().min(1, "Please select a plan"),
     joiningReason: z.string().optional(),
 });
-
-function getPlanDescription(plan: PaymentPlan, currencySymbol: string): string {
-    if (!plan) {
-        return "N/A";
-    }
-
-    switch (plan.type) {
-        case paymentPlanType.FREE:
-            return "Free plan";
-        case paymentPlanType.ONE_TIME:
-            return `One-time payment of ${currencySymbol}${plan.oneTimeAmount?.toFixed(2)}`;
-        case paymentPlanType.SUBSCRIPTION:
-            if (plan.subscriptionYearlyAmount) {
-                return `Billed annually at ${currencySymbol}${plan.subscriptionYearlyAmount.toFixed(2)}`;
-            }
-            return `${currencySymbol}${plan.subscriptionMonthlyAmount?.toFixed(2)} per month`;
-        case paymentPlanType.EMI:
-            return `${currencySymbol}${plan.emiAmount?.toFixed(2)} per month for ${plan.emiTotalInstallments} months`;
-        default:
-            return "N/A";
-    }
-}
-
-function getIncludedProductsDescription(
-    plan: PaymentPlan,
-    includedProducts: Course[],
-): Course[] {
-    if (!plan) {
-        return [];
-    }
-
-    if (plan.includedProducts && plan.includedProducts.length > 0) {
-        const includedProductsList = includedProducts.filter((product) =>
-            plan.includedProducts?.includes(product.courseId),
-        );
-        return includedProductsList;
-    }
-
-    return [];
-}
 
 export default function Checkout({
     product,
@@ -516,7 +122,7 @@ export default function Checkout({
             }
         };
 
-        if (profile.userId) {
+        if (profile?.userId) {
             fetchMembership();
         }
     }, [profile]);
@@ -576,7 +182,7 @@ export default function Checkout({
                         name: product.name,
                         image: product.featuredImage || siteinfo.logo?.file,
                         prefill: {
-                            email: profile.email,
+                            email: profile?.email || "",
                         },
                         handler: function (response) {
                             verifySignature(response);
@@ -711,7 +317,7 @@ export default function Checkout({
                     isOrderSummaryOpen={isOrderSummaryOpen}
                     setIsOrderSummaryOpen={setIsOrderSummaryOpen}
                 />
-                <div className="w-full grid md:grid-cols-[1fr,400px] gap-8 items-start">
+                <div className="w-full grid md:grid-cols-[1fr,400px] gap-8">
                     <div className="space-y-8">
                         {membershipStatus ===
                             Constants.MembershipStatus.ACTIVE ||
@@ -857,16 +463,6 @@ export default function Checkout({
                                                                                 theme={
                                                                                     theme
                                                                                 }
-                                                                                onSelect={(
-                                                                                    planId,
-                                                                                ) => {
-                                                                                    field.onChange(
-                                                                                        planId,
-                                                                                    );
-                                                                                    handlePlanSelection(
-                                                                                        planId,
-                                                                                    );
-                                                                                }}
                                                                             />
                                                                         );
                                                                     },
@@ -931,14 +527,12 @@ export default function Checkout({
                             </>
                         )}
                     </div>
-                    <div className="hidden md:block">
-                        <DesktopOrderSummary
-                            product={product}
-                            selectedPlan={selectedPlan}
-                            currencySymbol={currencySymbol}
-                            theme={theme}
-                        />
-                    </div>
+                    <DesktopOrderSummary
+                        product={product}
+                        selectedPlan={selectedPlan}
+                        currencySymbol={currencySymbol}
+                        theme={theme}
+                    />
                 </div>
             </div>
             <Script src="https://checkout.razorpay.com/v1/checkout.js" />
