+++ packages/common-logic/src/models/membership.ts
@@ -19,7 +19,7 @@ export const MembershipSchema = new mongoose.Schema<InternalMembership>(
             default: generateUniqueId,
         },
         userId: { type: String, required: true },
-        paymentPlanId: String,
+        paymentPlanId: { type: String, required: true },
         entityId: { type: String, required: true },
         entityType: {
             type: String,
@@ -27,6 +27,7 @@ export const MembershipSchema = new mongoose.Schema<InternalMembership>(
             required: true,
         },
         sessionId: { type: String, required: true, default: generateUniqueId },
+        isIncludedInPlan: { type: Boolean, default: false },
         status: {
             type: String,
             enum: Object.values(MembershipStatus),
@@ -40,7 +41,6 @@ export const MembershipSchema = new mongoose.Schema<InternalMembership>(
         rejectionReason: { type: String },
         subscriptionId: { type: String },
         subscriptionMethod: { type: String },
-        isIncludedInPlan: Boolean,
     },
     {
         timestamps: true,
